requirements.txt
# EcoFlow Cloud MQTT Publisher - Standalone Requirements
paho-mqtt>=1.6.0
aiohttp>=3.8.0
protobuf>=4.0.0
attrs>=21.0.0
jsonpath-ng>=1.5.0